Added few exceptions and functions

+++ Day4/LISP_gradle/app/src/main/java/org/factory/NodeFactory.java
@@ -9,7 +9,7 @@ import org.visitor.SymbolNode;
 
 public class NodeFactory {
 
-    public NumberNode createNumber(int value) {
+    public NumberNode createNumber(double value) {
         return new NumberNode(value);
     }
